+++ demo/src/main/java/com/demo/repository/GerenteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.demo.model.Gerente;
 
-public interface GerenteRepository extends JpaRepository<Gerente, Long> {
+public interface GerenteRepository extends JpaRepository<Gerente, Integer> {
 
 }
