+++ demo/src/main/java/com/demo/repository/ProveedorRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.demo.model.Proveedor;
 
-public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {
+public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {
 	
 
 }
